Move zerotier provider/resource logic to pkg
This doesn't really matter; but it is expected by the community in a lot
of instances. This arbitrary, and is more about smoothing over patch
adoption than anything else.
