+++ main.go
@@ -3,8 +3,7 @@ package main
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
-
-	"github.com/someara/terraform-provider-zerotier/zerotier"
+	"github.com/someara/terraform-provider-zerotier/pkg/zerotier"
 )
 
 func main() {
