+++ pkg/zerotier/resource_network.go
